Commuting_certification/dev/backend/requirements.txt
Django==2.2.9
djangorestframework==3.12.4
Pillow==8.2.0
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
